re-adding fps counter
